+++ sources/apex_plugin_template.sql
@@ -15,11 +15,11 @@ whenever sqlerror exit sql.sqlcode rollback
 --------------------------------------------------------------------------------
 begin
   wwv_flow_api.import_begin (
-    p_version_yyyy_mm_dd     => '2016.08.24'      ,
-    p_release                => '5.1.4.00.08'     ,
-    p_default_workspace_id   => 100000            ,
-    p_default_application_id => 100               ,
-    p_default_owner          => 'PLAYGROUND_DATA' );
+    p_version_yyyy_mm_dd     => '2016.08.24'  ,
+    p_release                => '5.1.4.00.08' ,
+    p_default_workspace_id   => 100000        ,
+    p_default_application_id => 100           ,
+    p_default_owner          => 'PLAYGROUND'  );
 end;
 /